Create Create Trace & 1 Free Surface Layer Synthetic Signal
Create a Trace class that has properties required of every trace.

Create a Synthetic class which is a subclass of Trace that is a synthetic signal for some model. In this case it is just a 1 layer model with a free surface. Therefore we model the primary and all its multiples.

The synthetic then returns the amplitudes of the primary and multiples that would have been recorded with this model setup.  This involves interpolating the amplitudes at theoretical arrival times to actual recording times.

Finally the theoretical amplitudes and arrival times are plotted in blue and the recorded amplitudes at recorded times are plotted in green.